requirements.txt
streamlit
mistralai
python-dotenv
pybase64
Pillow